Extract profile-related handlers into separate module
